First half of pause / resume forwarding
