+++ client/src/session_layer.rs
@@ -52,7 +52,7 @@ pub struct SessionLayerState {
     nodes_addr: HashMap<SocketAddr, NodeId>,
 
     forward_unreliable: HashMap<NodeId, ForwardSender>,
-    forward_paused_till: Option<DateTime<Utc>>,
+    forward_paused_till: Arc<RwLock<Option<DateTime<Utc>>>>,
 
     p2p_sessions: HashMap<NodeId, Arc<Session>>,
     ingress_channel: Channel<Forwarded>,
@@ -72,7 +72,7 @@ impl SessionsLayer {
                 sessions: Default::default(),
                 nodes_addr: Default::default(),
                 forward_unreliable: Default::default(),
-                forward_paused_till: Default::default(),
+                forward_paused_till: Arc::new(RwLock::new(Default::default())),
                 p2p_sessions: Default::default(),
                 ingress_channel: ingress,
                 starting_sessions: None,
@@ -300,8 +300,8 @@ impl SessionsLayer {
         // TODO: Use `_session` parameter. We can allow using other session, than default.
 
         let node = self.registry.resolve_node(node_id).await?;
-        let (sender, disconnected) = self.virtual_tcp.connect(node).await?;
-
+        let paused_check = { self.state.read().await.forward_paused_till.clone() };
+        let (sender, disconnected) = self.virtual_tcp.connect(node, paused_check).await?;
 
         //sender.override_send_with = self.get_next_fwd_payload()
 
@@ -337,30 +337,18 @@ impl SessionsLayer {
         Ok(tx)
     }
 
-    async fn get_next_fwd_payload<T>(&self, rx: &mut mpsc::Receiver<T>) -> Option<T> {
-        let raw_date = {
-            // dont lock the state longer then needed.
-            self.state.read().await.forward_paused_till
-        };
-        if let Some(date) = raw_date {
-            if let Ok(duration) = (date - Utc::now()).to_std() {
-                log::debug!("Receiver Paused!!! {:?}", duration);
-                tokio::time::delay_for(duration).await;
-                log::debug!("Receiver Continues...");
-            }
-            self.state.write().await.forward_paused_till = None;
-            log::debug!("reset date");
-        }
-        rx.next().await
-    }
-
     async fn forward_unreliable_handler(
         self,
         session: Arc<Session>,
         node: NodeEntry,
         mut rx: mpsc::Receiver<Vec<u8>>,
     ) {
-        while let Some(payload) = rx.next().await {
+        let paused_check = { self.state.read().await.forward_paused_till.clone() };
+        while let Some(payload) = self
+            .virtual_tcp
+            .get_next_fwd_payload(&mut rx, paused_check.clone())
+            .await
+        {
             log::trace!("Forwarding message (U) to {}", node.id);
 
             let forward = Forward::unreliable(session.id, node.slot, payload);
